+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * `ref0`, `ref1`, `ref2`, ... is now also usable outside of `GrammarDefinition`.
   * Use `resolve` to inline all the referenced parsers, which now also works with `SettableParser`.
   * Deprecated `removeSettables`, that is superseded by the more powerful `resolve` operation.
+* Add the possibility to join multiple `Token` and transform their values.
 
 ## 4.1.0